Ajuste valores de int para decimal

+++ InfinitePay/Responses/InstallmentsDetailResponse.cs
@@ -2,11 +2,11 @@
 
 public class InstallmentsDetailResponse
 {
-    public int amount { get; set; }
+    public decimal amount { get; set; }
     public double mdr_fee { get; set; }
-    public double anticipation_fee { get; set; }
+    public decimal anticipation_fee { get; set; }
     public DateTime payment_date { get; set; }
     public DateTime predicted_payment_date { get; set; }
-    public double predicted_anticipation_fee { get; set; }
+    public decimal predicted_anticipation_fee { get; set; }
 
 }
